+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/Configuration.java
@@ -8,11 +8,13 @@ public class Configuration implements Serializable {
     private final Path logDir;
     private final Path payloadDir;
     private final String agentJavaCmd;
+    private final Path dependencyCacheDir;
 
-    public Configuration(Path logDir, Path payloadDir, String agentJavaCmd) {
+    public Configuration(Path logDir, Path payloadDir, String agentJavaCmd, Path dependencyCacheDir) {
         this.logDir = logDir;
         this.payloadDir = payloadDir;
         this.agentJavaCmd = agentJavaCmd;
+        this.dependencyCacheDir = dependencyCacheDir;
     }
 
     public Path getLogDir() {
@@ -25,5 +27,9 @@ public class Configuration implements Serializable {
 
     public String getAgentJavaCmd() {
         return agentJavaCmd;
+    }
+
+    public Path getDependencyCacheDir() {
+        return dependencyCacheDir;
     }
 }